+++ packages/expo-observe/src/index.ts
@@ -2,7 +2,6 @@ import ExpoObserve from './module';
 
 export { default as AppMetrics } from 'expo-app-metrics';
 export { ObserveRoot } from './ObserveRoot';
-export { useObserve } from './useObserve';
 
 ExpoObserve.setBundleDefaults({
   environment: process.env.NODE_ENV ?? 'production',
@@ -11,3 +10,4 @@ ExpoObserve.setBundleDefaults({
 
 export { default } from './module';
 export * from './types';
+export { useObserve } from './useObserve';